+++ Version News.R
@@ -1,6 +1,0 @@
-
-
-
-News page.
-
-textProjectionPlot plot words extreme in several ways (before it only ploted words extreme in one such as frequency or p-value or dpp).