+++ NEWS.md
@@ -2,12 +2,12 @@
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
 
-# text 0.9.01
+# text 0.9.02 
 
 ### Major changes
-
+- Changing DEFAULT of cross-validation method in textTrainRegression and textTrainRandomForrest to 10-fold cross-validation in the inner loop as well. (0.9.02)
 ### Bug Fixes
-- `textProjectionPlot` plots words extreme in more than just one feature (i.e., words are now plotted that satisfy, for example, *both* `plot_n_word_extreme` and `plot_n_word_frequency`).
+- `textProjectionPlot` plots words extreme in more than just one feature (i.e., words are now plotted that satisfy, for example, *both* `plot_n_word_extreme` and `plot_n_word_frequency`). (0.9.01)
 
 ### Minor changes